+++ src/mruby_integration/core/timing.cpp
@@ -2,16 +2,6 @@
 #include "raylib.h"
 
 auto
-mrb_set_target_fps(mrb_state* mrb, mrb_value) -> mrb_value
-{
-  mrb_int fps;
-  mrb_get_args(mrb, "i", &fps);
-
-  SetTargetFPS(fps);
-  return mrb_nil_value();
-}
-
-auto
 mrb_get_fps(mrb_state* mrb, mrb_value) -> mrb_value
 {
   return mrb_int_value(mrb, GetFPS());
@@ -31,11 +21,6 @@ mrb_get_time(mrb_state* mrb, mrb_value) -> mrb_value
 void
 append_core_timing(mrb_state* mrb)
 {
-  mrb_define_method(mrb,
-                    mrb->kernel_module,
-                    "set_target_fps",
-                    mrb_set_target_fps,
-                    MRB_ARGS_REQ(1));
   mrb_define_method(
     mrb, mrb->kernel_module, "get_fps", mrb_get_fps, MRB_ARGS_NONE());
   mrb_define_method(mrb,